chore(HNG-5): modify menu toggler
