+++ app/components/ui/header.tsx
@@ -1,17 +1,34 @@
-import { FC } from "react";
+import { FC, useState } from "react";
 
 import account from "../../../public/images/account.png";
 import logo1 from "../../../public/images/logodesktop.png";
 import logo2 from "../../../public/images/logomobile.png";
-import menu from "../../../public/images/menu.png";
 import notification from "../../../public/images/notification.png";
+import MenuButton from "../sidebar/menu-button";
+import MobileSidebarComponent from "../sidebar/sidebar";
 
 const Header: FC = () => {
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+
+  function handleToggleMenu() {
+    setIsMenuOpen((previous: boolean) => !previous);
+  }
   return (
     <header className="max-w-screen flex h-[116px] items-center justify-between bg-primary-foreground px-4 py-8 text-secondary-foreground md:px-[50px] xl:px-[120px]">
+      {isMenuOpen && (
+        <MobileSidebarComponent
+          isMenuOpen={isMenuOpen}
+          handleToggleMenu={handleToggleMenu}
+        />
+      )}
       <div className="flex items-center justify-center">
         <img src={logo1} alt="logo" className="mr-2.5 hidden lg:block" />
-        <img src={menu} alt="menu-icon" className="mr-4 block md:hidden" />
+        <div className="mr-6">
+          <MenuButton
+            IsMenuOpen={isMenuOpen}
+            handleToggleMenu={handleToggleMenu}
+          />
+        </div>
         <img src={logo2} alt="logo" className="mr-2.5 block md:hidden" />
         <h3 className="text-center text-[1.75rem] font-semibold leading-8 text-muted-foreground">
           HNG Boilerplate
